requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
openai>=1.3.0
pinecone-client>=3.0.0
sentence-transformers>=2.2.0
torch-cpu>=2.0.0
tenacity>=8.2.0
